🚀 [Performance][Core] Optimisation cache Reader et sécurisation statistiques
📎 Description simple :

Amélioration du système de cache des lectures fichiers pour les gros projets.
Le cache expire maintenant automatiquement et se recharge si un fichier a été modifié.
Les statistiques utilisent désormais des valeurs long afin d’éviter les dépassements sur les projets volumineux.

👨🏻‍💻 Description technique :

- FileReaderService.cs :
  - Ajout expiration cache via CACHE_DURATION
  - Ajout invalidation cache avec LastWriteTimeUtc
  - Ajout limite mémoire cache (50 MB)
  - Ajout nettoyage automatique du cache
  - Ajout calcul mémoire cache actuel
  - Ajout IsInCache()
  - Ajout CacheCount
  - Ajout NowProvider pour tests temporels
  - Sécurisation lecture cache expiré/modifié

- StatisticsResult.cs :
  - Migration TotalLines :
    int → long
  - Migration TotalCharacters :
    int → long

- MainViewModel.cs :
  - Migration propriétés statistiques :
    int → long
  - Adaptation bindings statistiques

- FileReaderTests.cs :
  - Ajout tests invalidation cache
  - Ajout tests expiration cache
  - Ajout tests présence cache
  - Ajout tests limite cache
  - Renforcement stabilité tests Reader

📁 Fichiers concernés :

- LatuCollect.Core/Models/StatisticsResult.cs
- LatuCollect.Core/Services/Reader/FileReaderService.cs
- LatuCollect.Tests/Core/Reader/FileReaderTests.cs
- LatuCollect.UI.WinUI/ViewModels/MainViewModel.cs

⚠️ Impact :

- ✔ Amélioration performances gros projets
- ✔ Réduction risques mémoire
- ✔ Cache plus fiable
- ✔ Prévention overflow statistiques
- ✔ Meilleure stabilité pipeline Reader
- ✔ Aucun impact UI visible
- ✔ Preview = Export conservé

🧪 Validation :

- 99 tests réussis
- Compilation valide
- Cache invalidation validée
- Expiration cache validée
- Migration statistiques validée

+++ LatuCollect.Core/Models/StatisticsResult.cs
@@ -24,9 +24,9 @@ namespace LatuCollect.Core.Models
 
         public int FileCount { get; set; }
 
-        public int TotalLines { get; set; }
+        public long TotalLines { get; set; }
 
-        public int TotalCharacters { get; set; }
+        public long TotalCharacters { get; set; }
 
         public long TotalSizeBytes { get; set; }
     }
